Demonstrate bug when dumping UrlGenerator lazily
This demonstrates the original issue in symfony/symfony#2335 and also one issue described by @barros001 in symfony/symfony#2139. The situation involves defining a route whose pattern matches the prefix of a previously imported resource (we have this with the "/foo" import and "foo_index" route in routing.yml).

When the UrlGenerator is generated along with UrlMatcher on a proper cache warm-up, there is no problem:

        $ app/console cache:clear --env=prod
        $ curl http://localhost/app.php/foo/

However, when the cache is only cleared and allowed to warm up on the first request, the "foo_index" route is omitted from the dumped UrlGenerator:

        $ app/console cache:clear --no-warmup --env=prod
        $ curl http://localhost/app.php/foo/generate